+++ src/TGUI/ComboBox.cpp
@@ -296,7 +296,7 @@ namespace tgui
     sf::Vector2f ComboBox::getSize() const
     {
         if (m_Loaded)
-            return sf::Vector2f(m_ListBox->getSize().x, m_ListBox->getItemHeight() + m_TopBorder + m_BottomBorder);
+            return sf::Vector2f(m_ListBox->getSize().x, static_cast<float>(m_ListBox->getItemHeight() + m_TopBorder + m_BottomBorder));
         else
             return sf::Vector2f(0, 0);
     }
@@ -308,7 +308,7 @@ namespace tgui
         m_NrOfItemsToDisplay = nrOfItemsInList;
 
         if (m_NrOfItemsToDisplay < m_ListBox->m_Items.size())
-            m_ListBox->setSize(m_ListBox->getSize().x, (m_NrOfItemsToDisplay * m_ListBox->getItemHeight()) + 2*m_BottomBorder);
+            m_ListBox->setSize(m_ListBox->getSize().x, (m_NrOfItemsToDisplay * m_ListBox->getItemHeight()) + 2.0f*m_BottomBorder);
     }
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -431,7 +431,7 @@ namespace tgui
 
         // There is a minimum width
         if (m_ListBox->getSize().x < 50 + m_LeftBorder + m_RightBorder + m_TextureArrowDownNormal.getSize().x)
-            m_ListBox->setSize(50 + m_LeftBorder + m_RightBorder + m_TextureArrowDownNormal.getSize().x, m_ListBox->getSize().y);
+            m_ListBox->setSize(50.0f + m_LeftBorder + m_RightBorder + m_TextureArrowDownNormal.getSize().x, m_ListBox->getSize().y);
 
         // The item height needs to change
         m_ListBox->setItemHeight(itemHeight);
@@ -447,7 +447,7 @@ namespace tgui
 
         // Make room to add another item, until there are enough items
         if ((m_NrOfItemsToDisplay == 0) || (m_NrOfItemsToDisplay > m_ListBox->getItems().size()))
-            m_ListBox->setSize(m_ListBox->getSize().x, (m_ListBox->getItemHeight() * (m_ListBox->getItems().size() + 1)) + 2*m_BottomBorder);
+            m_ListBox->setSize(m_ListBox->getSize().x, (m_ListBox->getItemHeight() * (m_ListBox->getItems().size() + 1)) + 2.0f*m_BottomBorder);
 
         // Add the item
         return m_ListBox->addItem(item);
@@ -848,7 +848,7 @@ namespace tgui
         sf::Vector2f viewPosition = (target.getView().getSize() / 2.f) - target.getView().getCenter();
 
         // Get the global position
-        sf::Vector2f topLeftPosition = states.transform.transformPoint(getPosition() + sf::Vector2f(m_LeftBorder, m_TopBorder) + viewPosition);
+        sf::Vector2f topLeftPosition = states.transform.transformPoint(getPosition() + sf::Vector2f(static_cast<float>(m_LeftBorder), static_cast<float>(m_TopBorder)) + viewPosition);
         sf::Vector2f bottomRightPosition = states.transform.transformPoint(getPosition().x + m_ListBox->getSize().x - m_RightBorder - (m_TextureArrowDownNormal.getSize().x * (static_cast<float>(m_ListBox->getItemHeight()) / m_TextureArrowDownNormal.getSize().y)) + viewPosition.x,
                                                                        getPosition().y + m_ListBox->getSize().y - m_BottomBorder + viewPosition.y);
 
@@ -859,22 +859,22 @@ namespace tgui
         sf::Transform oldTransform = states.transform;
 
         // Draw left border
-        sf::RectangleShape border(sf::Vector2f(m_LeftBorder, m_ListBox->getItemHeight() + m_TopBorder + m_BottomBorder));
+        sf::RectangleShape border(sf::Vector2f(static_cast<float>(m_LeftBorder), static_cast<float>(m_ListBox->getItemHeight() + m_TopBorder + m_BottomBorder)));
         border.setFillColor(m_ListBox->m_BorderColor);
         target.draw(border, states);
 
         // Draw top border
-        border.setSize(sf::Vector2f(m_ListBox->getSize().x, m_TopBorder));
+        border.setSize(sf::Vector2f(m_ListBox->getSize().x, static_cast<float>(m_TopBorder)));
         target.draw(border, states);
 
         // Draw right border
         border.setPosition(m_ListBox->getSize().x - m_RightBorder, 0);
-        border.setSize(sf::Vector2f(m_RightBorder, m_ListBox->getItemHeight() + m_TopBorder + m_BottomBorder));
+        border.setSize(sf::Vector2f(static_cast<float>(m_RightBorder), static_cast<float>(m_ListBox->getItemHeight() + m_TopBorder + m_BottomBorder)));
         target.draw(border, states);
 
         // Draw bottom border
-        border.setPosition(0, m_ListBox->getItemHeight() + m_TopBorder);
-        border.setSize(sf::Vector2f(m_ListBox->getSize().x, m_BottomBorder));
+        border.setPosition(0, static_cast<float>(m_ListBox->getItemHeight() + m_TopBorder));
+        border.setSize(sf::Vector2f(m_ListBox->getSize().x, static_cast<float>(m_BottomBorder)));
         target.draw(border, states);
 
         // Move the front rect a little bit
