+++ src/TGUI/Slider2d.cpp
@@ -486,8 +486,8 @@ namespace tgui
                         std::string::size_type commaPos = value.find(',');
                         if ((commaPos != std::string::npos) && (value.find(',', commaPos) == std::string::npos))
                         {
-                            setMinimum(sf::Vector2f(atof(value.substr(1, commaPos-1).c_str()),
-                                                    atof(value.substr(commaPos+1, value.length()-commaPos-2).c_str())));
+                            setMinimum(sf::Vector2f(static_cast<float>(atof(value.substr(1, commaPos-1).c_str())),
+                                                    static_cast<float>(atof(value.substr(commaPos+1, value.length()-commaPos-2).c_str()))));
                         }
                         else
                             TGUI_OUTPUT("TGUI error: Failed to parse 'Minimum' property.");
@@ -507,8 +507,8 @@ namespace tgui
                         std::string::size_type commaPos = value.find(',');
                         if ((commaPos != std::string::npos) && (value.find(',', commaPos) == std::string::npos))
                         {
-                            setMaximum(sf::Vector2f(atof(value.substr(1, commaPos-1).c_str()),
-                                                    atof(value.substr(commaPos+1, value.length()-commaPos-2).c_str())));
+                            setMaximum(sf::Vector2f(static_cast<float>(atof(value.substr(1, commaPos-1).c_str())),
+                                                    static_cast<float>(atof(value.substr(commaPos+1, value.length()-commaPos-2).c_str()))));
                         }
                         else
                             TGUI_OUTPUT("TGUI error: Failed to parse 'Maximum' property.");
@@ -528,8 +528,8 @@ namespace tgui
                         std::string::size_type commaPos = value.find(',');
                         if ((commaPos != std::string::npos) && (value.find(',', commaPos) == std::string::npos))
                         {
-                            setValue(sf::Vector2f(atof(value.substr(1, commaPos-1).c_str()),
-                                                  atof(value.substr(commaPos+1, value.length()-commaPos-2).c_str())));
+                            setValue(sf::Vector2f(static_cast<float>(atof(value.substr(1, commaPos-1).c_str())),
+                                                  static_cast<float>(atof(value.substr(commaPos+1, value.length()-commaPos-2).c_str()))));
                         }
                         else
                             TGUI_OUTPUT("TGUI error: Failed to parse 'Value' property.");
